resolving conflicts in database dir
